+++ src/constants/const/index.ts
@@ -1,0 +1,1 @@
+export { sizeMapping } from "./mapping";
