+++ src/constants/const/mapping.ts
@@ -1,0 +1,9 @@
+export const sizeMapping = {
+  tiny: "Tiny",
+  small: "Small",
+  medium: "Medium",
+  large: "Large",
+  "large-x": "LargeX",
+  "large-xx": "Large2X",
+  "large-xxx": "Large3X",
+}
