+++ src/constants/enums/index.ts
@@ -1,0 +1,1 @@
+export { EAlign } from "./alignment";
