--- src/constants/types.ts
@@ -1,7 +1,0 @@
-enum Alignment {
-	Left = 0,
-	Right = 1,
-	Center = 2,
-}
-
-export default { Alignment };
