+++ src/constants/types/common.ts
@@ -1,7 +1,3 @@
-// Size
-export type Size = "tiny" | "small" | "medium" | "large" | "large-x" | "large-xx" | "large-xxx";
-// Size plus
-export type SizePlus = "xs" | "sm" | "md" | "lg" | "xl" | "xxl" | "xxxl" | "o-xxxl" | "xo-xxxl";
 // Content Direction
 export type ContentDirection = "ltr" | "rtl" | "auto";
 // Container Layout
@@ -23,11 +19,6 @@ export type HeadingSize = "h1" | "h2" | "h3" | "h4" | "h5" | "h6";
 
 // Grid TypeType
 export type GridType = "default" | "masonry";
-
-// List type
-export type List = "unordered" | "ordered" | "description";
-export type UnorderedListMarker = "none" | "disc" | "circle" | "square";
-export type OrderedListType = "1" | "A" | "a" | "I" | "i";
 
 // Media type
 export type MediaType = "image" | "video" | "audio";
