+++ src/constants/types/gradient.ts
