+++ src/utils/index.ts
@@ -1,19 +1,1 @@
 export { isObject, isStringOrStringArray } from "./function";
-export {
-	ContentDirection,
-	ContainerLayout,
-	DisplayZone,
-	GradientLinearDir,
-	GradientLinearDirAngles,
-	GradientLinearDirDiagonal,
-	GradientType,
-	ObjectLayout,
-	ProfileHidden,
-	RadialGradientShapeSize,
-	Size,
-	SizePlus,
-	List,
-	MessageType,
-	UnorderedListMarker,
-	OrderedListType,
-} from "./styles";
