--- src/utils/styles/index.ts
@@ -1,23 +1,0 @@
-export {
-	ContentDirection,
-	ContainerLayout,
-	DisplayZone,
-	HeadingSize,
-	MediaType,
-	ObjectLayout,
-	ProfileHidden,
-	Ratio,
-	Size,
-	SizePlus,
-	List,
-	MessageType,
-	UnorderedListMarker,
-	OrderedListType,
-} from "./common";
-export {
-	GradientLinearDir,
-	GradientLinearDirAngles,
-	GradientLinearDirDiagonal,
-	GradientType,
-	RadialGradientShapeSize,
-} from "./gradient";
